pyproject.toml
[project]
name = "python-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "webdriver-manager>=4.0.2",
    "selenium>=4.25.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
]
